+++ NEWS.md
@@ -1,5 +1,6 @@
 # httr2 (development version)
 
+* `curl_transform()` will now use `req_body_json_modify()` for JSON data (#258).
 * New `url_modify()` makes it easier to modify an existing url (#464).
 * New `req_url_relative()` for constructing relative urls (#449).
 * `url_parse()` gains `base_url` argument so you can also use it to parse relative URLs (#449).